pyproject.toml
[project]
name = "podcasticot"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.2",
    "fastapi[standard]>=0.117.1",
    "feedparser==6.0.11",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=0.19.2",
    "requests>=2.27.1",
]
[dependency-groups]
dev = [
    "pytest==8.2.0",
    "ruff>=0.13.2",
]
